+++ node-graph/gcore/src/vector/algorithms/mod.rs
@@ -1,3 +1,4 @@
 mod instance;
 mod merge_by_distance;
 pub mod offset_subpath;
+pub mod position_on_bezpath;
